text_lambda/requirements.txt
psycopg2-binary
boto3
pandas
python-dotenv